Novice level
Script that can encode words based on a fixed shift. The amount and direction of the shift is up to you (shift of +3 means that an “A” becomes a “D”). The encode script asks the user for a word or sentence and then print the results of the encoded text.